+++ packages/core/src/index.ts
@@ -1,0 +1,11 @@
+import { Component } from './Component';
+import { MetaF } from './JSX';
+import * as renderer from './renderer';
+import { State } from './State';
+
+export {
+    renderer,
+    State,
+    Component,
+};
+export default MetaF;