+++ packages/core/src/sync/sync.ts
@@ -15,9 +15,9 @@ const methodsToThrow = [
     'construct',
 ];
 export class PromiseCache {
-    content = new WeakMap<Promise<any>, any>();
-    proxies = new WeakMap<Promise<any>, any>();
-    resolved = Promise.resolve<any>(undefined);
+    content = new WeakMap<Promise<unknown>, unknown>();
+    proxies = new WeakMap<Promise<unknown>, unknown>();
+    resolved = Promise.resolve<unknown>(undefined);
 }
 const promiseCache = new PromiseCache();
 const trap = () => {
@@ -25,14 +25,14 @@ const trap = () => {
 };
 const pendingHandler = methodsToThrow.reduce((result, method) => ({...result,  [method]: trap}), {});
 type Primitives = string | number | boolean | symbol | undefined | null;
-export type PrimitivesWrapper<T extends any> = T extends Primitives ? { value: T } : T;
-export type AsyncPrimitivesWrapper<T extends any> = T extends Promise<infer R>
+export type PrimitivesWrapper<T> = T extends Primitives ? { value: T } : T;
+export type AsyncPrimitivesWrapper<T> = T extends Promise<infer R>
     ? PrimitivesWrapper<R>
     : T;
 export function sync<P>(promise: P): AsyncPrimitivesWrapper<P> {
     if (!(promise instanceof Promise)) return promise as AsyncPrimitivesWrapper<P>;
-    if (promiseCache.content.has(promise)) return promiseCache.content.get(promise);
-    if (promiseCache.proxies.has(promise)) return promiseCache.proxies.get(promise);
+    if (promiseCache.content.has(promise)) return promiseCache.content.get(promise) as AsyncPrimitivesWrapper<P>;
+    if (promiseCache.proxies.has(promise)) return promiseCache.proxies.get(promise) as AsyncPrimitivesWrapper<P>;
     promiseCache.resolved = Promise.all([
         promiseCache.resolved,
         promise.then(value => promiseCache.content.set(promise, wrapPrimitive(value))),
@@ -43,7 +43,7 @@ export function sync<P>(promise: P): AsyncPrimitivesWrapper<P> {
     return pendingProxy;
 }
 
-function wrapPrimitive<T extends any>(value: T) {
+function wrapPrimitive<T extends unknown>(value: T): PrimitivesWrapper<T> {
     if (typeof value === 'string' ||
         typeof value === 'number' ||
         typeof value === 'boolean' ||