+++ packages/core/src/utils/index.ts
@@ -1,5 +1,0 @@
-export function ensure<T>(value: T | undefined): T {
-    if (value === undefined) throw new Error('Value provided is undefined');
-
-    return value;
-}