+++ tests/game/src/Opponent/index.tsx
@@ -1,0 +1,8 @@
+import { Component } from 'metaf-core';
+import './style.less';
+
+export class Opponent extends Component() {
+    render() {
+        return (1);
+    }
+}