+++ tests/game/src/index.ts
@@ -1,0 +1,4 @@
+import { renderer } from 'metaf-core';
+import { App } from './App';
+
+renderer.renderToDOM(App, document.getElementById('root')!);